memleaks patch from ubuntu/debian, original source unknown.
This was a non-issue in original source as the reverb engine was disabled due to unresolved issues at time of release but someone downstream re-enabled it.

This patch may be redundant since the reverb engine in wildmidi has been rebuilt and will be added to and enabled in this version.

git-svn-id: svn://svn.code.sf.net/p/wildmidi/svn/trunk@46 c49d981e-1564-41db-a9da-c6d570d03c1e