AMIGAOS4: Fix compiler error with Seek()
Seek() is a dos.library function and deprecated in favour of ChangeFilePosition().

In the latest SDK the obsolete status has been forced by changing the name to OBSOLETESeek().

A proper fix would be to rewrite to use the fseek()/lseek() pair or aforementioned ChangeFilePosition().

A quick workaround is to
#include <dos/obsolete.h>
into
#ifdef __amigaos4__
to compile a working binary again.

Which it does and which I propose.

Thank you(cherry picked from commit a2bc3538e7b40e5e6eda667a826369ffc129dbaa)