Replace `__FUNCTION__` with C99's `__func__`.
`__FUNCTION__` is a non-standard compiler extension. `__func__` is
the standard equivalent.

WildMIDI's codebase is largely written in C89, but a C99-ism or two
has crept in over time, such as C++-style comments.

It might be possible to eliminate these at some point. And have the
entire project be C89-compatible.