Fix broken MUS drum channel.
As described in #234, if the same MUS is played twice from the same
memory buffer, then channels 9 and 15 (the drum channel) will be
swapped.

I've addressed this by removing the modification of the memory buffer
entirely, to prevent future issues of this type.

Fixes https://github.com/Mindwerks/wildmidi/issues/234