Silence warnings when compiling as ANSI C.
`__func__` is the standard C equivalent to `__FUNCTION__`.

Compiling WildMIDI as ANSI C with GCC warnings enabled causes many
warning messages to be printed which complain about `__FUNCTION__`
not existing in C89. This commit works around it by only using it
when compiling as C99 or newer. In C89, the function string is simply
dummied-out instead.

While I was at it, I used some macro magic to hide most uses of
`__FUNCTION__` and `__LINE__`.