Discard extraneous 'end of track' events in Type 1 MIDI files.
WildMIDI internally uses a single list of events, like a Type 0 MIDI
file, resulting in Type 1 MIDI files being 'flattened' into a single
track when they are loaded.

Because of this, 'end of track' events lose their original meaning,
ending the entire song instead of just the track which they belong
to.

Rather than overhaul WildMIDI to internally use multiple tracks, it
is easier to simply discard all but the final 'end of track' event.

https://www.doomworld.com/idgames/music/ultimidi

This fixes the playback of 'E2M1 TrialD - Pagliacci's Going to Town',
which would otherwise loop within seconds of starting.

Closes #213. The MIDI files from Rise of the Triad which are linked
in the issue are Type 1 MIDI files which suffer from this exact
problem.